chore: Fix the github page redirect issue
